Rename queryinput package to api
