+++ API/src/main/java/de/brandwatch/minianalytics/api/postgres/repository/QueryRepository.java
@@ -1,6 +1,6 @@
-package de.brandwatch.minianalytics.queryinput.postgres.repository;
+package de.brandwatch.minianalytics.api.postgres.repository;
 
-import de.brandwatch.minianalytics.queryinput.postgres.model.Query;
+import de.brandwatch.minianalytics.api.postgres.model.Query;
 import org.springframework.data.jpa.repository.JpaRepository;
